feat: Implement endpoint to get total number of jobs
Added a new endpoint `getTotalJobs` to retrieve the total count of jobs from the database. This endpoint uses the `Job` model to count the total number of documents in the jobs collection. If successful, it returns a JSON response with the total number of jobs. In case of an error, it returns a 500 Internal Server Error response.

This commit implements error handl