Add nodemailer and passport dependencies
In this commit, we add two new dependencies to the project:

- nodemailer: "^6.9.13"
- passport: "^0.7.0"

Nodemailer is a module for Node.js applications to allow easy email sending. This will be used for handling email functionalities within our application.

Passport is a popular authentication middleware for Node.js. With this dependency, we'll be able to easily implement authentication strategies in our application.

These dependencies are essential for the upcoming features that require email functionality and user authentication.